// src/sdk/typegen/Registry.ts
export type IdentityInput = Enum<{ Address: AddressInput, ContractId: ContractIdInput }>;
export type IdentityOutput = Enum<{ Address: AddressOutput, ContractId: ContractIdOutput }>;
export enum InitializationErrorInput { CannotReinitialized = 'CannotReinitialized' };
export enum InitializationErrorOutput { CannotReinitialized = 'CannotReinitialized' };
export type MetadataInput = Enum<{ B256: [], Bytes: Bytes, Int: BigNumberish, String: StdString }>;
export type MetadataOutput = Enum<{ B256: [], Bytes: Bytes, Int: BN, String: StdString }>;
// ...
export type AddressInput = { bits: string };
export type AddressOutput = AddressInput;
export type AssetIdInput = { bits: string };
export type AssetIdOutput = AssetIdInput;
export type ContractIdInput = { bits: string };
export type ContractIdOutput = ContractIdInput;
// ...
const storageSlots: StorageSlot[] = [
  {
    "key": "93b67ee4f0f76b71456fb4385c86aec15689e1ce5f6f6ac63b71716afa052998",
    "value": "0000000000000000000000000000000000000000000000000000000000000000"
  }
];

export class RegistryInterface extends Interface {
  constructor() {
    super(abi);
  }

  declare functions: {
    decimals: FunctionFragment;
    name: FunctionFragment;
    symbol: FunctionFragment;
    total_assets: FunctionFragment;
    total_supply: FunctionFragment;
    metadata: FunctionFragment;
    domain_exists: FunctionFragment;
    get_domain_asset_id: FunctionFragment;
    get_domain_name: FunctionFragment;
    get_expiration: FunctionFragment;
    get_expiration_by_parent: FunctionFragment;
    get_grace_period: FunctionFragment;
    get_resolver: FunctionFragment;
    initialize: FunctionFragment;
    is_domain_active: FunctionFragment;
    register_high_level_domain: FunctionFragment;
    register_sub_domain: FunctionFragment;
    renew_domain: FunctionFragment;
    resolve_to_primary_domain: FunctionFragment;
    set_primary: FunctionFragment;
    set_resolver: FunctionFragment;
  };
}

export class Registry extends Contract {
  static readonly abi = abi;
  static readonly storageSlots = storageSlots;

  declare interface: RegistryInterface;
  declare functions: {
    decimals: InvokeFunction<[asset: AssetIdInput], Option<number>>;
    name: InvokeFunction<[asset: AssetIdInput], Option<StdString>>;
    symbol: InvokeFunction<[asset: AssetIdInput], Option<StdString>>;
    total_assets: InvokeFunction<[], BN>;
    total_supply: InvokeFunction<[asset: AssetIdInput], Option<BN>>;
    metadata: InvokeFunction<[asset: AssetIdInput, key: StdString], Option<MetadataOutput>>;
    domain_exists: InvokeFunction<[asset_id: AssetIdInput], boolean>;
    get_domain_asset_id: InvokeFunction<[domain: StdString], AssetIdOutput>;
    get_domain_name: InvokeFunction<[asset: AssetIdInput], StdString>;
    get_expiration: InvokeFunction<[domain: StdString], Option<BN>>;
    get_expiration_by_parent: InvokeFunction<[name: StdString, parent: StdString], Option<BN>>;
    get_grace_period: InvokeFunction<[domain: StdString], Option<BN>>;
    get_resolver: InvokeFunction<[domain: StdString], Option<ContractIdOutput>>;
    initialize: InvokeFunction<[], void>;
    is_domain_active: InvokeFunction<[asset_id: AssetIdInput], boolean>;
    register_high_level_domain: InvokeFunction<[recipient: IdentityInput, name: StdString], AssetIdOutput>;
    register_sub_domain: InvokeFunction<[recipient: IdentityInput, parent: StdString, name: StdString, expiration: Option<BigNumberish>, grace_period: Option<BigNumberish>, resolver: ContractIdInput], AssetIdOutput>;
    renew_domain: InvokeFunction<[name: StdString, parent: StdString, expiration: BigNumberish], void>;
    resolve_to_primary_domain: InvokeFunction<[identity: IdentityInput], Option<AssetIdOutput>>;
    set_primary: InvokeFunction<[asset: AssetIdInput], void>;
    set_resolver: InvokeFunction<[domain: StdString, resolver: ContractIdInput], void>;
  };

  constructor(
    id: string | AbstractAddress,
    accountOrProvider: Account | Provider,
  ) {
    super(id, abi, accountOrProvider);
  }
}